[Fix] Compiler warning
- Type conversion in scripting file fixed

+++ Engine/core-engine/src/Scripting/scripting-system.cpp
@@ -60,7 +60,7 @@ namespace Copium::Utils
 		if (!stream) return nullptr;
 		std::streampos end = stream.tellg();
 		stream.seekg(0, std::ios::beg);
-		uint32_t size = end - stream.tellg();
+		uint32_t size = (uint32_t)(end - stream.tellg());
 		if (size == 0) return nullptr;
 		char* buffer = new char[size];
 		stream.read((char*)buffer, size);
